AI/requirement.txt
fastapi
uvicorn
sqlmodel
pytest
httpx
langchain
langchain-community
ollama
python-dotenv